Move getStructuredQuestionUrlWithParameters out of metabase-lib/v1 and into metabase/dashboard, including related helpers
